FEAT: 좋아요/싫어요 바텀시트 구현 (카톡링크제외)

+++ app/src/main/java/org/appcenter/inudorm/presentation/mypage/MateListActivity.kt
@@ -22,7 +22,7 @@ abstract class MateListActivity : AppCompatActivity() {
         recyclerViewToSetup: RecyclerView,
         radioButton1: RadioButton,
         radioButton2: RadioButton,
-        onDetailOpen: () -> Unit
+        onDetailOpen: (Int) -> Unit
     ) {
         mRadioButton1 = radioButton1
         mRadioButton2 = radioButton2
